Revisit Unity support
Unconditionally use a header bar, and check XDG_CURRENT_DESKTOP to
decide whether to use it as the title bar. Use gtk-dialogs-use-header
to decide whether to use a header bar on the preferences dialog.

Zero UI changes for GNOME users. It looks better (due to the more
reasonable icon size) for Unity users.